Rename notepad module file
