+++ src/components/item/container.js
@@ -2,7 +2,7 @@ import React from 'react'
 import { connect } from 'react-redux'
 import { Resizable } from '../resizable'
 import { EsperContainer } from '../esper'
-import { NotePad } from '../note'
+import { NotePad } from '../note/index.js'
 import * as act from '../../actions'
 import cx from 'classnames'
 import { SASS, ITEM } from '../../constants'
