+++ src/components/note/index.js
@@ -1,4 +1,4 @@
-export * from './list'
-export * from './list-item'
-export * from './pad'
-export * from './toolbar'
+export * from './list.js'
+export * from './list-item.js'
+export * from './notepad.js'
+export * from './toolbar.js'
