+++ src/components/note/notepad.js
